Fixed Review issues
